All beautiful buttons

+++ app/controllers/home_controller.rb
@@ -44,22 +44,26 @@ class HomeController < ApplicationController
 
   def after_load
     song = params[:song]
-    File.open(Rails.root.join('public', 'music', song.original_filename.gsub(' ', '_')), 'wb') do |file|
-      file.write(song.read)
-    end
-    new_song = Song.new(name: params[:song].original_filename.delete(".mp3"))
-    singers = params[:singer].gsub(' ft. ', ', ').split(', ')
-    singers.each do |singer|
-      check_singer = Singer.find_by_name(singer)
-      if check_singer.nil?
-        new_singer = Singer.create!(name:singer)
-        new_singer.songs << new_song
-      else
-        check_singer.songs << new_song
+    unless song.nil? || params[:singer] == ""
+      File.open(Rails.root.join('public', 'music', song.original_filename.gsub(' ', '_')), 'wb') do |file|
+        file.write(song.read)
       end
+      new_song = Song.new(name: params[:song].original_filename.delete(".mp3"))
+      singers = params[:singer].gsub(' ft. ', ', ').split(', ')
+      singers.each do |singer|
+        check_singer = Singer.find_by_name(singer)
+        if check_singer.nil?
+          new_singer = Singer.create!(name:singer)
+          new_singer.songs << new_song
+        else
+          check_singer.songs << new_song
+        end
+      end
+      new_song.save
+      redirect_to load_path, notice: 'Загрузка выполнена успешно'
+    else
+      redirect_to load_path, notice: 'Вы не заполнили все условия'
     end
-    new_song.save
-    redirect_to load_path, notice: 'Загрузка выполнена успешно'
   end
 
   def creating_playlist
